# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ORBIT ERP API",
        "version": "1.0.0"
    },
    "paths": {
        "/budgetHeads": {
            "get": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "Returns all budgetHeads on the system",
                "description": "Returns all budgetHeads on the system",
                "operationId": "findRoles",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetHead"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "Save user budgetHead",
                "operationId": "budgetHeads",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/budgetHeads/{id}": {
            "get": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "Get budgetHead by id",
                "description": "Returns based on id ",
                "operationId": "showRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "budgetHead id to get",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetHead"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid budgetHead id"
                    }
                }
            },
            "put": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "update budgetHead by database",
                "description": "Updates budgetHead in database",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "budgetHead id to update",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BudgetHead updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetHead"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid budgetHead id"
                    }
                }
            },
            "delete": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "remove budgetHead from database",
                "description": "Deletes budgetHead in database",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "budgetHead id to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BudgetHead deleted successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetHead"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid budgetHead id"
                    }
                }
            }
        },
        "/budgetHeads/{id}/edit": {
            "get": {
                "tags": [
                    "BudgetHeads"
                ],
                "summary": "Open form to edit budgetHead",
                "description": "Returns based on id ",
                "operationId": "editRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "budgetHead id to edit",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetHead"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid budgetHead id"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BudgetHead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "05"
                    },
                    "code": {
                        "type": "string",
                        "example": "BSG/34/21"
                    },
                    "name": {
                        "type": "string",
                        "example": "ICT Budget"
                    },
                    "description": {
                        "type": "string",
                        "example": "Budget Description"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2020-10-20"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2020-12-22"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BudgetHead"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}